requirements.txt
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
Django==2.2.1
djangorestframework==3.9.4
freeze==1.0.10
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
MarkupSafe==1.1.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
Pygments==2.4.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
